appcode/requirements.txt
fastapi
uvicorn[standard]
openai
google-cloud-firestore
PyYAML
